Rework classes a bit, make quantity private for encapsulation

+++ OperationOOP.Core/Models/Soda.cs
@@ -10,7 +10,7 @@ namespace OperationOOP.Core.Models;
 public class Soda : Beverage
 {
     public bool IsSugarFree { get; set; }
-    public Soda(int id, string name, double volume, decimal price, int quantity, bool isSugarFree)
+    public Soda(int id, string name, decimal volume, decimal price, int quantity, bool isSugarFree)
            : base(id, name, volume, price, quantity)
     {
